Move compatibility backports to separate file
